+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Implemented `.to_bits()` and `.from_bits()` for `APyCFloat`.
 - LaTeX representation for `APyCFixed` and `APyCFloat`.
+- Array comparisons
 
 ### Fixed